src/main/webapp/app/entities/aide/route/aide-routing-resolve.service.spec.ts
jest.mock('@angular/router');

import { TestBed } from '@angular/core/testing';
import { HttpResponse } from '@angular/common/http';
import { HttpClientTestingModule } from '@angular/common/http/testing';
import { ActivatedRouteSnapshot, Router } from '@angular/router';
import { of } from 'rxjs';

import { IAide, Aide } from '../aide.model';
import { AideService } from '../service/aide.service';

import { AideRoutingResolveService } from './aide-routing-resolve.service';

describe('Aide routing resolve service', () => {
  let mockRouter: Router;
  let mockActivatedRouteSnapshot: ActivatedRouteSnapshot;
  let routingResolveService: AideRoutingResolveService;
  let service: AideService;
  let resultAide: IAide | undefined;

  beforeEach(() => {
    TestBed.configureTestingModule({
      imports: [HttpClientTestingModule],
      providers: [Router, ActivatedRouteSnapshot],
    });
    mockRouter = TestBed.inject(Router);
    mockActivatedRouteSnapshot = TestBed.inject(ActivatedRouteSnapshot);
    routingResolveService = TestBed.inject(AideRoutingResolveService);
    service = TestBed.inject(AideService);
    resultAide = undefined;
  });

  describe('resolve', () => {
    it('should return IAide returned by find', () => {
      // GIVEN
      service.find = jest.fn(id => of(new HttpResponse({ body: { id } })));
      mockActivatedRouteSnapshot.params = { id: 123 };

      // WHEN
      routingResolveService.resolve(mockActivatedRouteSnapshot).subscribe(result => {
        resultAide = result;
      });

      // THEN
      expect(service.find).toBeCalledWith(123);
      expect(resultAide).toEqual({ id: 123 });
    });

    it('should return new IAide if id is not provided', () => {
      // GIVEN
      service.find = jest.fn();
      mockActivatedRouteSnapshot.params = {};

      // WHEN
      routingResolveService.resolve(mockActivatedRouteSnapshot).subscribe(result => {
        resultAide = result;
      });

      // THEN
      expect(service.find).not.toBeCalled();
      expect(resultAide).toEqual(new Aide());
    });

    it('should route to 404 page if data not found in server', () => {
      // GIVEN
      jest.spyOn(service, 'find').mockReturnValue(of(new HttpResponse({ body: null as unknown as Aide })));
      mockActivatedRouteSnapshot.params = { id: 123 };

      // WHEN
      routingResolveService.resolve(mockActivatedRouteSnapshot).subscribe(result => {
        resultAide = result;
      });

      // THEN
      expect(service.find).toBeCalledWith(123);
      expect(resultAide).toEqual(undefined);
      expect(mockRouter.navigate).toHaveBeenCalledWith(['404']);
    });
  });
});
